+++ vendors/adapters/iqm2_adapter_async.py
@@ -26,7 +26,7 @@ from datetime import datetime, timedelta
 from urllib.parse import urljoin
 import aiohttp
 from vendors.adapters.base_adapter_async import AsyncBaseAdapter, logger
-from vendors.utils.item_filters import should_skip_procedural_item
+from pipeline.filters import should_skip_item
 from bs4 import BeautifulSoup
 
 
@@ -217,7 +217,7 @@ class AsyncIQM2Adapter(AsyncBaseAdapter):
         items_before = len(items)
         items = [
             item for item in items
-            if not should_skip_procedural_item(item.get('title', ''))
+            if not should_skip_item(item.get('title', ''))
         ]
         items_filtered = items_before - len(items)
         if items_filtered > 0: